WinderDeploy/app/KoreanDiscordAssistant/requirementscloud.txt
discord.py>=2.5.2
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
pytz>=2024.1
requests>=2.32.3